pyproject.toml
[project]
name = "mem0-server"
version = "0.1.0"
description = "Mem0 MCP Server with Ollama embeddings and Anthropic LLM"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.26.0",
    "fastmcp>=2.14.5",
    "mem0ai>=1.0.3",
    "anthropic>=0.79.0",
    "ollama>=0.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
